Style: 포스트 정보 컴포넌트 분리(조회수, 좋아요 개수, 댓글 개수)
