+++ src/components/Board/Board/Post.jsx
@@ -1,12 +1,9 @@
 import React from "react";
 import { Chip, Typography } from "@material-tailwind/react";
 import { useSearchParams } from "react-router-dom";
-import { MdOutlineRemoveRedEye, MdThumbUp } from "react-icons/md";
-import { BsFillChatDotsFill } from "react-icons/bs";
 import Tags from "./Tags";
-import { getDateInfo } from "../../../utils/date";
 import { BOARD_CATEGORIE_WITHOUT_TOGGLE_TYPES } from "../../../constants/category";
-import useBreakpoint from "../../../hooks/useBreakPoint";
+import PostInfo from "./PostInfo";
 
 export default function Post({
   post: {
@@ -22,7 +19,6 @@ export default function Post({
     commentCount,
   },
 }) {
-  const { isSmallMobile } = useBreakpoint();
   const [searchParams] = useSearchParams();
   const type = searchParams.get("type");
 
@@ -45,34 +41,14 @@ export default function Post({
         {content}
       </Typography>
       <Tags tags={tags} />
-      <div className="flex justify-between items-center gap-4">
-        <div className="flex items-center gap-2">
-          <Typography variant="small" className="text-white font-medium">
-            {author.name}
-          </Typography>
-          {!isSmallMobile && (
-            <Typography variant="small" className="text-white font-medium">
-              {updatedAt
-                ? getDateInfo(new Date(updatedAt))
-                : getDateInfo(new Date(createdAt))}
-            </Typography>
-          )}
-        </div>
-        <ul className="flex items-center gap-4">
-          <li className="flex items-center gap-2 text-white font-medium">
-            <MdOutlineRemoveRedEye />
-            <Typography className="mt-1">{viewCount}</Typography>
-          </li>
-          <li className="flex items-center gap-2 text-white font-medium">
-            <MdThumbUp />
-            <Typography className="mt-1">{likeCount}</Typography>
-          </li>
-          <li className="flex items-center gap-2 text-white font-medium">
-            <BsFillChatDotsFill />
-            <Typography className="mt-1">{commentCount}</Typography>
-          </li>
-        </ul>
-      </div>
+      <PostInfo
+        author={author}
+        createdAt={createdAt}
+        updatedAt={updatedAt}
+        viewCount={viewCount}
+        likeCount={likeCount}
+        commentCount={commentCount}
+      />
     </li>
   );
 }
